feat(web): mobile-first meeting actions (a11y, nav, bundle)
Committed items from the mobile-first design review:

1. Promote Barkley to primary mobile tab bar, demote Rewards
   (nav.ts). Adds nav.barkleyShort to both locales. Barkley is the
   in-the-moment reward loop and now reachable in one tap from any
   route.
2. MobileTabBar: bump label size from text-[11px] to text-xs (12 px),
   widen icon/label gap, add aria-label on each tab link. Add a
   sr-only aria-live="polite" region in the authenticated shell so
   TalkBack and VoiceOver announce the active page on route changes.
3. Lazy-load the dashboard route. Extracts DashboardPage into
   dashboard-page.tsx and imports it via React.lazy + Suspense in
   the route file. motion/react (~40 kB gzip) and the dashboard
   sub-tree now ship as a separate 47 kB gzip chunk loaded only
   when /dashboard is visited.
4. Dialog close button: size=icon (44 px on mobile, 32 px on desktop)
   to meet WCAG 2.5.5 target size; aria-label via i18n common.close
   ("Fermer" / "Close") replaces the sr-only span; move inset to
   top-4 right-4 on mobile for a less muscle-memory-hostile position.
5. Symptom form conflict banner pairs the status-warning color with
   an AlertTriangle icon and role="alert" so color-blind and low-
   vision users get the signal without relying on hue (WCAG 1.4.1).
