+++ apps/web/src/lib/i18n/locales/en.json
@@ -5,6 +5,7 @@
     "loading": "Loading...",
     "save": "Save",
     "cancel": "Cancel",
+    "close": "Close",
     "back": "Back",
     "retry": "Retry",
     "language": "Language",
@@ -204,6 +205,7 @@
     "journal": "Journal",
     "medications": "Medications",
     "barkley": "Barkley program",
+    "barkleyShort": "Barkley",
     "news": "News",
     "account": "My account",
     "home": "Home",
