+++ apps/web/src/lib/i18n/locales/fr.json
@@ -5,6 +5,7 @@
     "loading": "Chargement...",
     "save": "Enregistrer",
     "cancel": "Annuler",
+    "close": "Fermer",
     "back": "Retour",
     "retry": "Réessayer",
     "language": "Langue",
@@ -204,6 +205,7 @@
     "journal": "Journal",
     "medications": "Médicaments",
     "barkley": "Programme Barkley",
+    "barkleyShort": "Barkley",
     "news": "Actualités",
     "account": "Mon compte",
     "home": "Accueil",
